Changed chain policy from a hard-coded list to a dynamic module import.
