+++ Abe/Chain/CryptoCash.py
@@ -1,0 +1,28 @@
+# Copyright(C) 2014 by Abe developers.
+
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as
+# published by the Free Software Foundation, either version 3 of the
+# License, or (at your option) any later version.
+# 
+# This program is distributed in the hope that it will be useful, but
+# WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+# Affero General Public License for more details.
+# 
+# You should have received a copy of the GNU Affero General Public
+# License along with this program.  If not, see
+# <http://www.gnu.org/licenses/agpl.html>.
+
+from .NvcChain import NvcChain
+
+class CryptoCash(NvcChain):
+    def __init__(chain, **kwargs):
+        chain.name = 'Cash'
+        chain.code3 = 'CAS'
+        chain.address_version = "\x22"
+        chain.magic = "\xe4\xc6\xfe\xe7"
+        NvcChain.__init__(chain, **kwargs)
+
+    datadir_conf_file_name = "Cash.conf"
+    datadir_rpcport = 3941
