+++ Abe/Chain/__init__.py
@@ -14,26 +14,13 @@
 # License along with this program.  If not, see
 # <http://www.gnu.org/licenses/agpl.html>.
 
-import deserialize
-from deserialize import opcodes
-import BCDataStream
-import util
+from .. import deserialize, BCDataStream, util
+from ..deserialize import opcodes
 
 def create(policy, **kwargs):
-    # XXX It's about time to interpret policy as a module name.
-    if policy in [None, "Bitcoin"]: return Bitcoin(**kwargs)
-    if policy == "Testnet":         return Testnet(**kwargs)
-    if policy == "Namecoin":        return Namecoin(**kwargs)
-    if policy == "LegacyNoBit8":    return Sha256Chain(**kwargs)
-    if policy == "NovaCoin":        return NovaCoin(**kwargs)
-    if policy == "CryptoCash":      return CryptoCash(**kwargs)
-    if policy == "Hirocoin":        return Hirocoin(**kwargs)
-    if policy == "X11":             return X11Chain(**kwargs)
-    if policy == "X11Pos":          return X11PosChain(**kwargs)
-    if policy == "Bitleu":          return Bitleu(**kwargs)
-    if policy == "Keccak":          return KeccakChain(**kwargs)
-    if policy == "Maxcoin":         return Maxcoin(**kwargs)
-    return Sha256NmcAuxPowChain(**kwargs)
+    mod = __import__(__name__ + '.' + policy, fromlist=[policy])
+    cls = getattr(mod, policy)
+    return cls(policy=policy, **kwargs)
 
 
 PUBKEY_HASH_LENGTH = 20
@@ -62,10 +49,10 @@ SCRIPT_TYPE_MULTISIG = 5
 SCRIPT_TYPE_P2SH = 6
 
 
-class Chain(object):
+class BaseChain(object):
     def __init__(chain, src=None, **kwargs):
         for attr in [
-            'id', 'magic', 'name', 'code3', 'address_version', 'decimals', 'script_addr_vers']:
+            'id', 'policy', 'magic', 'name', 'code3', 'address_version', 'decimals', 'script_addr_vers']:
 
             if attr in kwargs:
                 val = kwargs.get(attr)
@@ -235,187 +222,3 @@ class Chain(object):
 
     datadir_conf_file_name = "bitcoin.conf"
     datadir_rpcport = 8332
-
-class Sha256Chain(Chain):
-    def block_header_hash(chain, header):
-        return util.double_sha256(header)
-
-class Bitcoin(Sha256Chain):
-    def __init__(chain, **kwargs):
-        chain.name = 'Bitcoin'
-        chain.code3 = 'BTC'
-        chain.address_version = '\x00'
-        chain.script_addr_vers = '\x05'
-        chain.magic = '\xf9\xbe\xb4\xd9'
-        Chain.__init__(chain, **kwargs)
-
-class Testnet(Sha256Chain):
-    def __init__(chain, **kwargs):
-        chain.name = 'Testnet'
-        chain.code3 = 'BC0'
-        chain.address_version = '\x6f'
-        chain.script_addr_vers = '\xc4'
-        chain.magic = '\xfa\xbf\xb5\xda'
-        Chain.__init__(chain, **kwargs)
-
-    # XXX
-    #datadir_conf_file_name = "bitcoin.conf"
-    #datadir_rpcport = 8332
-
-class NmcAuxPowChain(Chain):
-    def __init__(chain, **kwargs):
-        chain.block_version_bit_merge_mine = 8
-        Chain.__init__(chain, **kwargs)
-
-    def ds_parse_block_header(chain, ds):
-        d = Chain.ds_parse_block_header(chain, ds)
-        if d['version'] & (1 << 8):
-            d['auxpow'] = deserialize.parse_AuxPow(ds)
-        return d
-
-    def has_feature(chain, feature):
-        return feature == 'block_version_bit8_merge_mine'
-
-class Sha256NmcAuxPowChain(Sha256Chain, NmcAuxPowChain):
-    pass
-
-class Namecoin(Sha256NmcAuxPowChain):
-    def __init__(chain, **kwargs):
-        chain.name = 'Namecoin'
-        chain.code3 = 'NMC'
-        chain.address_version = '\x34'
-        chain.magic = '\xf9\xbe\xb4\xfe'
-        Chain.__init__(chain, **kwargs)
-
-    _drops = (opcodes.OP_NOP, opcodes.OP_DROP, opcodes.OP_2DROP)
-
-    def parse_decoded_txout_script(chain, decoded):
-        start = 0
-        pushed = 0
-
-        # Tolerate (but ignore for now) name operations.
-        for i in xrange(len(decoded)):
-            opcode = decoded[i][0]
-
-            if decoded[i][1] is not None or \
-                    opcode == opcodes.OP_0 or \
-                    opcode == opcodes.OP_1NEGATE or \
-                    (opcode >= opcodes.OP_1 and opcode <= opcodes.OP_16):
-                pushed += 1
-            elif opcode in chain._drops:
-                to_drop = chain._drops.index(opcode)
-                if pushed < to_drop:
-                    break
-                pushed -= to_drop
-                start = i + 1
-            else:
-                return Chain.parse_decoded_txout_script(chain, decoded[start:])
-
-        return SCRIPT_TYPE_UNKNOWN, decoded
-
-
-    datadir_conf_file_name = "namecoin.conf"
-    datadir_rpcport = 8336
-
-class LtcScryptChain(Chain):
-    def block_header_hash(chain, header):
-        import ltc_scrypt
-        return ltc_scrypt.getPoWHash(header)
-
-class PpcPosChain(Chain):
-    def ds_parse_transaction(chain, ds):
-        return deserialize.parse_Transaction(ds, has_nTime=True)
-
-    def ds_parse_block(chain, ds):
-        d = Chain.ds_parse_block(chain, ds)
-        d['block_sig'] = ds.read_bytes(ds.read_compact_size())
-        return d
-
-class NvcChain(LtcScryptChain, PpcPosChain):
-    def has_feature(chain, feature):
-        return feature == 'nvc_proof_of_stake'
-
-class NovaCoin(NvcChain):
-    def __init__(chain, **kwargs):
-        chain.name = 'NovaCoin'
-        chain.code3 = 'NVC'
-        chain.address_version = "\x08"
-        chain.magic = "\xe4\xe8\xe9\xe5"
-        chain.decimals = 6
-        Chain.__init__(chain, **kwargs)
-
-    datadir_conf_file_name = "novacoin.conf"
-    datadir_rpcport = 8344
-
-class CryptoCash(NvcChain):
-    def __init__(chain, **kwargs):
-        chain.name = 'Cash'
-        chain.code3 = 'CAS'
-        chain.address_version = "\x22"
-        chain.magic = "\xe4\xc6\xfe\xe7"
-        Chain.__init__(chain, **kwargs)
-
-    datadir_conf_file_name = "Cash.conf"
-    datadir_rpcport = 3941
-
-class X11Chain(Chain):
-    def block_header_hash(chain, header):
-        import xcoin_hash
-        return xcoin_hash.getPoWHash(header)
-
-class X11PosChain(X11Chain, PpcPosChain):
-    pass
-
-class Hirocoin(X11Chain):
-    def __init__(chain, **kwargs):
-        chain.name = 'Hirocoin'
-        chain.code3 = 'HIRO'
-        chain.address_version = '\x28'
-        chain.script_addr_vers = '\x05'
-        chain.magic = '\xfe\xc4\xb9\xde'
-        Chain.__init__(chain, **kwargs)
-
-    datadir_conf_file_name = 'hirocoin.conf'
-    datadir_rpcport = 9347
-    datadir_p2pport = 9348
-
-YAC_START_TIME = 1377557832
-
-class ScryptJaneChain(Chain):
-    def block_header_hash(chain, header):
-        import yac_scrypt
-        b = chain.parse_block_header(header)
-        return yac_scrypt.getPoWHash(header, b['nTime'] + YAC_START_TIME - chain.start_time)
-
-class Bitleu(ScryptJaneChain, PpcPosChain):
-    def __init__(chain, **kwargs):
-        chain.name = 'Bitleu'
-        chain.code3 = 'BTL'
-        chain.address_version = "\x30"
-        chain.script_addr_vers = '\x1b'
-        chain.magic = "\xd9\xe6\xe7\xe5"
-        chain.decimals = 6
-        Chain.__init__(chain, **kwargs)
-
-    datadir_conf_file_name = "Bitleu.conf"
-    datadir_rpcport = 7997
-    start_time = 1394480376
-
-class KeccakChain(Chain):
-    def block_header_hash(chain, header):
-        return util.sha3_256(header)
-
-class Maxcoin(KeccakChain):
-    def __init__(chain, **kwargs):
-        chain.name = 'Maxcoin'
-        chain.code3 = 'MAX'
-        chain.address_version = '\x6e'
-        chain.script_addr_vers = '\x70'
-        chain.magic = "\xf9\xbe\xbb\xd2"
-        Chain.__init__(chain, **kwargs)
-
-    def transaction_hash(chain, binary_tx):
-        return util.sha256(binary_tx)
-
-    datadir_conf_file_name = 'maxcoin.conf'
-    datadir_rpcport = 8669
